admin/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
djangorestframework_simplejwt==5.0.0
mysqlclient==2.0.1
django-mysql==3.9
djongo==1.3.5
pymongo==3.12.1
django-cors-headers==3.5.0
drf_yasg==1.21.7